trying to make this an app

+++ wrapper.sh
@@ -8,26 +8,27 @@ function newVersion {
 	echo $NEW_VERSION
 }
 
-WEBHOOK="${1}"
+WEBHOOK="${webhookFile}"
 CLONE_URL=`jq '.repository.clone_url' $WEBHOOK | tr -d '"'`
 REPO_NAME=`basename ${CLONE_URL%%.git}`
 DESCRIPTION_FILE="$REPO_NAME/agave.json"
 
 REF=`jq '.ref' $WEBHOOK | tr -d '"'` 			# ref/tags/branch
-CHECK_BRANCH=`basename $REF`				# ref/tags/BRANCH
+BRANCH=`basename $REF`					# ref/tags/BRANCH
 CHECK_TAGS=`basename $(dirname $REF)`			# ref/TAGS/branch
-CHECK_CREATED=`jq '.created' $WEBHOOK | tr -d '"'`
+CREATED=`jq '.created' $WEBHOOK | tr -d '"'`
 
 IS_RELEASE=false
 if [ "$CHECK_TAGS" == "tags" ]; then
 	IS_RELEASE=true
-elif ! [ "$CHECK_BRANCH" == "master" ] && ! [ $CHECK_CREATED = false ]; then
+elif ! [ "$BRANCH" == "master" ] && ! [ $CREATED = false ]; then
 	echo "This is not a simple commit or release. Exiting without updating app."
 	exit
 fi
 
 # clone repo
-git clone $CLONE_URL
+#git clone $CLONE_URL
+git clone -b $BRANCH $CLONE_URL
 
 # check for app description
 if ! [ -e "$DESCRIPTION_FILE" ]; then
@@ -47,6 +48,15 @@ if [[ "$PREV_VERSION" == "(sourceref)" ]]; then
 	# update description file
         NEW_VERSION="${PREV_VERSION/(sourceref)/$REPLACEMENT}"
 	CHANGE_DESCRIPTION_FILE=`jq --arg foo $NEW_VERSION '.version = $foo' $DESCRIPTION_FILE`
+	rm $DESCRIPTION_FILE
+	echo $CHANGE_DESCRIPTION_FILE >> $DESCRIPTION_FILE
+fi
+
+# append branch name
+NAME=`jq '.name' $DESCRIPTION_FILE | tr -d '"'`
+if ! [ "$BRANCH" == "master" ] && ! [ "$NAME" == *"-$BRANCH" ]; then
+	NAME_REPLACEMENT="$NAME-$BRANCH"
+	CHANGE_DESCRIPTION_FILE=`jq --arg foo $NAME_REPLACEMENT '.name = $foo' $DESCRIPTION_FILE`
 	rm $DESCRIPTION_FILE
 	echo $CHANGE_DESCRIPTION_FILE >> $DESCRIPTION_FILE
 fi
